% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados_yt_charts.R
\name{dados_yt_charts}
\alias{dados_yt_charts}
\title{Extract YouTube Chart Data}
\usage{
dados_yt_charts(id, time_period)
}
\arguments{
\item{id}{Character vector containing the city codes.}

\item{time_period}{Character vector containing the time period for data extraction.}
}
\value{
A list with the response data from YouTube Charts API.
}
\description{
This function extracts data from the YouTube Charts for the specified city and time period.
}
\examples{
\dontrun{
  data <- dados_yt_charts(id = c("0x123", "0x456"), time_period = c("2023-09", "2023-10"))
}
}
