% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estados_alvo.R
\name{estados_alvo}
\alias{estados_alvo}
\title{Target States for YouTube Charts Data Extraction}
\usage{
estados_alvo()
}
\value{
A vector of state names or codes.
}
\description{
This function provides a list of target states for which YouTube charts data should be extracted.
}
\examples{
\dontrun{estados_alvo()}
}
