% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrai_artistas.R
\name{extrai_artistas}
\alias{extrai_artistas}
\title{Extract Artists from YouTube Charts Data}
\usage{
extrai_artistas(dados_yt_charts, alvo)
}
\arguments{
\item{data}{The dataset containing YouTube Charts data.}
}
\value{
A data frame with artist information.
}
\description{
This function extracts artist information from the YouTube Charts data.
}
\examples{
\dontrun{
  artist_data <- extrai_artistas(data)
}
}
