% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrai_id.R
\name{extrai_id}
\alias{extrai_id}
\title{Extract IDs from YouTube Charts Data}
\usage{
extrai_id(
  url_estados_cidades_br,
  time_period = c("2022-06-01/2023-06-30", "2023-06-01/2024-06-30")
)
}
\arguments{
\item{data}{The dataset containing YouTube Charts data.}
}
\value{
A vector with unique IDs.
}
\description{
This function extracts unique IDs from the YouTube Charts dataset.
}
\examples{
\dontrun{
  ids <- extrai_id(data)
}
}
