% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrai_musicas.R
\name{extrai_musicas}
\alias{extrai_musicas}
\title{Extract Songs from YouTube Charts Data}
\usage{
extrai_musicas(dados_yt_charts, alvo)
}
\arguments{
\item{data}{The dataset containing YouTube Charts data.}
}
\value{
A data frame with song information.
}
\description{
This function extracts song information from the YouTube Charts data.
}
\examples{
\dontrun{
  song_data <- extrai_musicas(data)
}
}
