% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipios_alvo.R
\name{municipios_alvo}
\alias{municipios_alvo}
\title{Target Municipalities for YouTube Charts Data Extraction}
\usage{
municipios_alvo(pop = 70000)
}
\value{
A vector of municipality names or codes.
}
\description{
This function provides a list of target municipalities for YouTube charts data extraction.
}
\examples{
\dontrun{municipios_alvo()}
}
