% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_estados_cidade_br.R
\name{url_estados_cidade_br}
\alias{url_estados_cidade_br}
\title{Generate URL for YouTube Charts by State and City}
\usage{
url_estados_cidade_br(alvo_query)
}
\arguments{
\item{state}{The state code.}

\item{city}{The city code.}
}
\value{
A URL string for the YouTube Charts API.
}
\description{
This function generates the URL used to extract YouTube Charts data for a specific state and city.
}
\examples{
\dontrun{
  url <- url_estados_cidade_br(state = "SP", city = "SaoPaulo")
}
}
